+++ crates/wash-lib/src/start/mod.rs
@@ -52,6 +52,26 @@
 //!     Ok(())
 //! }
 //! ```
+
+use anyhow::Result;
+pub async fn wait_for_server(url: &str, service: &str) -> Result<()> {
+    let mut wait_count = 1;
+    loop {
+        // Magic number: 10 + 1, since we are starting at 1 for humans
+        if wait_count >= 11 {
+            anyhow::bail!("Ran out of retries waiting for host to start");
+        }
+        match tokio::net::TcpStream::connect(url).await {
+            Ok(_) => break,
+            Err(e) => {
+                log::debug!("Waiting for {service} at {url} to come up, attempt {wait_count}. Will retry in 1 second. Got error {:?}", e);
+                wait_count += 1;
+                tokio::time::sleep(std::time::Duration::from_secs(1)).await;
+            }
+        }
+    }
+    Ok(())
+}
 mod nats;
 pub use nats::*;
 mod wasmcloud;
